requirements.txt
numpy
opencv
pillow
scikit-image
tqdm
pytorch 
torchvision
